requirements.txt
python-telegram-bot>=20.0
httpx>=0.24.0
python-dotenv>=1.0.0
python-telegram-bot[job-queue]